created <Media /> component that contains the audio and video for the Artist page
