+++ src/Styles/Utils/Artist.css
@@ -281,195 +281,6 @@
   color: var(--clr-light);
 }
 
-.artist-content {
-  margin: 20px 0px 10px 0px;
-}
-
-.audio-list {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-  margin: 20px 0px 0px 0px;
-  border-bottom: 1px solid var(--clr-news-break);
-}
-
-.audio-list-item {
-  padding: 8px;
-  border-top: 1px solid var(--clr-news-break);
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  cursor: pointer;
-}
-.audio-list-item.selected {
-  border: 1px solid var(--clr-accent);
-  background-color: var(--clr-accent-screen);
-  border-radius: 10px;
-}
-.audio-list-selected {
-  padding: 8px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  cursor: pointer;
-  border: 1px solid var(--clr-accent);
-  background-color: var(--clr-accent-screen);
-  border-radius: 10px;
-}
-
-.audio-list-art {
-  width: 50px;
-  border-radius: 10px;
-}
-
-.audio-list-text {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-}
-
-.audio-list-title-container {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: 100%;
-}
-
-.audio-list-title {
-  font-family: 'Titillium Web', sans-serif;
-  font-size: 1.15em;
-  font-weight: 600;
-  line-height: 1.2;
-  color: var(--clr-dark);
-  padding: 0px 10px;
-  text-align: left;
-  display: -webkit-box;
-  -webkit-line-clamp: 1;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.explicit-box {
-  border: 1px solid var(--clr-news-title);
-  padding: 0px 4px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  height: 18px;
-  margin: 0% 0px 0px 0px;
-}
-
-.explicit-text {
-  font-family: 'Titillium Web', sans-serif;
-  font-size: 0.7em;
-  font-weight: 600;
-  line-height: 0.9;
-  color: var(--clr-news-title);
-  transform: scaleY(95%);
-  transform: scaleX(110%);
-}
-
-.audio-list-subtext {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-}
-
-.audio-list-artist {
-  font-family: 'Titillium Web', sans-serif;
-  font-size: 0.95em;
-  font-weight: 400;
-  line-height: 1.15;
-  color: var(--clr-dark);
-  padding: 0px 0px 0px 10px;
-  text-align: left;
-  display: flex;
-  flex-direction: row;
-}
-
-.audio-list-album {
-  font-family: 'Merriweather', sans-serif;
-  font-size: 0.75em;
-  font-weight: 400;
-  line-height: 1.15;
-  font-style: italic;
-  color: var(--clr-text);
-  padding: 2px 0px 0px 10px;
-  text-align: left;
-}
-
-.audio-list-date {
-  font-family: 'Merriweather', sans-serif;
-  font-size: 0.7em;
-  font-weight: 400;
-  line-height: 1.15;
-  font-style: italic;
-  color: var(--clr-text);
-  padding: 2px 0px 0px 5px;
-  text-align: left;
-}
-
-.audio-list-item:hover .audio-list-art {
-  box-shadow: 0 4px 8px var(--clr-shadow2);
-}
-.audio-list-item:hover .audio-list-art,
-.audio-list-item:hover .audio-list-title,
-.audio-list-item:hover .audio-list-artist,
-.audio-list-item:hover .audio-list-album,
-.audio-list-item:hover .audio-list-date,
-.audio-list-item:hover .explicit-box {
-  transform: translateY(-2px);
-  transition: transform 0.15s ease-in-out;
-}
-.audio-list-item:active .audio-list-art {
-  box-shadow: 0 3px 6px var(--clr-shadow2);
-}
-.audio-list-item:active .audio-list-art,
-.audio-list-item:active .audio-list-title,
-.audio-list-item:active .audio-list-artist,
-.audio-list-item:active .audio-list-album,
-.audio-list-item:active .audio-list-date,
-.audio-list-item:hover .explicit-box {
-  transform: translateY(1px);
-  transition: transform 0.15s ease-in-out;
-}
-
-.audio-list-item:hover .audio-list-title,
-.audio-list-item:hover .audio-list-artist,
-.audio-list-item:hover .audio-list-album,
-.audio-list-item:hover .audio-list-date,
-.audio-list-selected:hover .audio-list-title,
-.audio-list-selected:hover .audio-list-artist,
-.audio-list-selected:hover .audio-list-album,
-.audio-list-selected:hover .audio-list-date {
-  color: var(--clr-accent);
-}
-
-.artist-video {
-  width: 100%;
-  max-width: 1000px;
-  height: auto;
-  margin: 0 auto;
-  display: block;
-}
-
-.video-title-artist-container {
-  width: 100%;
-  display: flex;
-  justify-content: left;
-}
-
-.video-title-artist {
-  font-family: 'Open Sans', sans-serif;
-  font-size: 1.15em;
-  font-weight: 400;
-  color: var(--clr-dark);
-  margin: 5px 0px 5px 5px;
-  text-align: left;
-}
-
 .artist-news {
   margin-top: 20px;
 }
